+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Recommendation: for ease of reading, use the following order:
 - Pin version for `aws-sdk-s3` crate version, that includes breaking changes
 - Update all minor versions of other crates
 - Cleanup duplicate image
+- Denormalization: `DatasetEntry` now contains a copy of owner's account name
+   for faster dataset handle resolutions without extra round trip to database
 
 ## [0.233.4] - 2025-04-22
 ### Changed